undetected_chromedriver 3.1.0 compatibility
